requirements.txt
absl-py==2.1.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
build==1.2.2
certifi==2024.8.30
charset-normalizer==3.3.2
chex==0.1.86
cloudpickle==3.1.0
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.3.0
crafter==1.8.3
cycler==0.12.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1727240639244/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dm-tree==0.1.8
-e git+https://github.com/penn-pal-lab/scaffolder.git@e9f26d9d7d3418379bd4e46282dc40ba875d8470#egg=dreamerv3
etils==1.5.2
exceptiongroup==1.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
Farama-Notifications==0.0.4
flatbuffers==24.3.25
FLORIS==2.4
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
glfw==2.7.0
google-pasta==0.2.0
grpcio==1.66.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
-e git+https://github.com/edwhu/Gymnasium-Robotics.git@0141b4a3fe160e7bf6a01afe3f440f4259e48b5c#egg=gymnasium_robotics
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jaxlib==0.4.30
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
ml-dtypes==0.4.1
mujoco==3.2.3
namex==0.0.8
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy==2.0.2
nvidia-cublas-cu12==12.6.1.4
nvidia-cuda-cupti-cu12==12.6.68
nvidia-cuda-nvcc-cu12==12.6.68
nvidia-cuda-runtime-cu12==12.6.68
nvidia-cudnn-cu12==9.4.0.58
nvidia-cufft-cu12==11.2.6.59
nvidia-cusolver-cu12==11.6.4.69
nvidia-cusparse-cu12==12.5.3.3
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.6.68
opensimplex==0.4.5.1
opt_einsum==3.4.0
optax==0.2.3
optree==0.12.1
packaging==24.1
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pettingzoo==1.24.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
pluggy==1.5.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
protobuf==4.25.5
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1725737916340/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyglet==1.5.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyOpenGL==3.1.7
pyparsing==3.2.0
pyproj==3.6.1
pyproject_hooks==1.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1725448933977/work
requests==2.32.3
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sklearn==0.0.post12
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.24.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.2
toolz==0.12.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724955920300/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.4
wind_farm_gym @ file:///home/maranc/Windfarm-Optimization-RL/wind-farm-env/dist/wind_farm_gym-0.0.3-py3-none-any.whl#sha256=72b5d07845c35d3da748162569c9bbe29c2a73c57627c56af7d4c7945a3d6f1b
wrapt==1.16.0
zipp==3.20.2
